+++ backend/data/ingested_data (prev) copy.csv
@@ -1,4 +1,4 @@
-post_image_urls
+post_image_urls, 
 https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/471997046_18618118558001321_7420361941174662615_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=Xx5qXRU8RawQ7kNvgEIGcxJ&_nc_gid=8ec73a7128e040c49787d1d09bc84afc&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYAHtZLGZmQwY4i65EMzedMMyy8VE2-Kq2F7StwRRh_apg&oe=677E7365&_nc_sid=d885a2
 https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/471989791_18597973315056421_2992110397002673459_n.jpg?stp=dst-jpg_e35_s1080x1080_sh0.08_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=m-4Gsp4zlsgQ7kNvgFLVJPB&_nc_gid=bc36bcdcf74845a09c4d22854164616d&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYBCOXDSsusnF2qGi8xiRxTSc2wN0PMVU-zMFdnH5iAZoA&oe=677E9972&_nc_sid=d885a2
 https://instagram.fdel27-3.fna.fbcdn.net/v/t39.30808-6/471199666_1243948443758360_3973335327799519382_n.jpg?stp=dst-jpg_e15_fr_s1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=lcpHHv4ds-IQ7kNvgFeI_fn&_nc_gid=8fec2a8774e142c68e6febf14a3ebee9&edm=ANTKIIoAAAAA&ccb=7-5&oh=00_AYAl0Bsc558YMZ5n6w2o37SqvhIRXD8upFjZgOLKEQL2BQ&oe=677E9407&_nc_sid=d885a2
