--- backend/data_ingestion_engine/ingested_data.csv
@@ -1,18 +1,0 @@
-,post_image_urls
-0,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/471997046_18618118558001321_7420361941174662615_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=VnJteccwjrwQ7kNvgFpr__9&_nc_gid=90db1da7e23f4a6ea1ad7ca4d8fec442&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYA_cuaYHL23wZj2_PspDshKlE0fMIW92qoK86q1ZFJigQ&oe=677EABA5&_nc_sid=d885a2
-1,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/471989791_18597973315056421_2992110397002673459_n.jpg?stp=dst-jpg_e35_s1080x1080_sh0.08_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=m-4Gsp4zlsgQ7kNvgGDI01f&_nc_gid=ff29174c8cfa4fb583a8f80c13e4a593&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYDF40E6mFeH4t3Uq2NeebiqAR1_C18FBm-DuqsHsVpBVQ&oe=677E9972&_nc_sid=d885a2
-2,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/472413874_3906408492933741_3337462021825972041_n.jpg?stp=dst-jpg_e35_s1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=uh7fCxswpb0Q7kNvgGUS2gu&_nc_gid=7a558d5f3a014b77bc2cee3027827b40&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYDKFaXSCjK-ngepxeJYFGVC_rE-gdpy8BZFzXeVHBJlMA&oe=677EB57E&_nc_sid=d885a2
-3,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/472442372_544157595334180_954069735060975616_n.jpg?stp=dst-jpg_e35_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=Pdp2RHQPsfoQ7kNvgGWZ60w&_nc_gid=433a8fa973f94121a4eab6ebf7be757b&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYCFewXUPOeeogWgo4RaxA7Vs8bwJvLAvKkdbJpb2Tgv8A&oe=677EB38A&_nc_sid=d885a2
-4,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/472280250_1794232208021285_3361994920465290213_n.jpg?stp=dst-jpg_e35_p1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=VRyVwG6PEfoQ7kNvgEGdK1m&_nc_gid=4416325ef8ad488d9111f266ecb98c0e&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYDmSGmntKxAlj2QtYpkA2Q2m3JCyB7eIVIxCUGLXBxJ2Q&oe=677ECBFD&_nc_sid=d885a2
-5,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/471853169_18443428489072786_6063938770476564880_n.jpg?stp=dst-jpg_e35_p1080x1080_sh0.08_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=8-FfaRiAAqUQ7kNvgGUNdXr&_nc_gid=7222b1c50e7b428e8a802db03410cff0&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYD_6YcUCd91yP5rRgDTzWhiqFHXo9tGeDQsB3WAI8KgYg&oe=677EB9E5&_nc_sid=d885a2
-6,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/472368800_18524787964000183_4604000888294897897_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=7914M2lbZlUQ7kNvgFAbL1z&_nc_gid=83e738d1d5e94a2e840cb8983729bff8&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYA0hk1jwxOyVi6j2tYdHhnTikTuUnPU6IhVyqvXjFpn0Q&oe=677EA26B&_nc_sid=d885a2
-7,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/472292196_18524758432000183_7031884123863766303_n.jpg?stp=dst-jpg_e35_p1080x1080_sh0.08_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=npXjIzi_sLwQ7kNvgE_8zKU&_nc_gid=bb0692867c03483bbf6f4383837e3616&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYAsnpJUP2UOY6i64BlFR8HaV0Oq53rHe9uBne8tt41gNg&oe=677E9F3D&_nc_sid=d885a2
-8,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/472387071_18524429929000183_6185800727945180494_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=2EMzlWO987oQ7kNvgGi0KTa&_nc_gid=727d6f1cfa714527b96bf05d16c0279d&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYB9vLFATR1IcCzkw5fYEPhFgf6OlXxl7RTTmpZur8mpAQ&oe=677E97CE&_nc_sid=d885a2
-9,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/468503039_1241725673797407_3858043650018235471_n.jpg?stp=dst-jpg_e35_s1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=MGvm6VPecvkQ7kNvgEOQ8-E&_nc_gid=0bc3d7857c464a239216dbe17f1fdd7e&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYCxBWzrS3tuTMnQuZygoV7ipVfVAF413tWb61lbWsLeew&oe=677EC9E6&_nc_sid=d885a2
-10,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/471527093_1586993615514161_7342768906940667684_n.jpg?stp=dst-jpg_e35_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=zzvzeviIi5MQ7kNvgEG2GHo&_nc_gid=8cf6df61930b41358bfaf13bcb6c4056&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYDg4h9GMsXT-IOuwtStq5wEWNaCsqmuGdIo4ErGF_SAqQ&oe=677ECDAC&_nc_sid=d885a2
-11,https://instagram.fdel27-6.fna.fbcdn.net/v/t51.2885-15/471614409_18527257438008035_7871872394580930984_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-6.fna.fbcdn.net&_nc_cat=111&_nc_ohc=_lbWfnUuCHkQ7kNvgHXqjzo&_nc_gid=a8d32d4a227a40bca2bc60ef162af45a&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYAozryCG8AqQDRAd1uG8udEtUS0mZCtY7ixur6CLUC1CA&oe=677ECF2A&_nc_sid=d885a2
-12,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/472528007_1667286207496561_3692269272234047369_n.jpg?stp=dst-jpg_e15_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=lwrpsPPtkS4Q7kNvgEjDPlO&_nc_gid=ac946b233644437ba47fc7a3ecca8e86&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYC_wpffKr4AMbr2I1CP-SLXH57pK-GNvQBrUg-ely0C1g&oe=677E9A7A&_nc_sid=d885a2
-13,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.2885-15/472297082_18574982497048326_1342415333300964666_n.jpg?stp=dst-jpg_e35_p1080x1080_sh0.08_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=O-JuqJIFpD4Q7kNvgFqjcy6&_nc_gid=06fd9061ea8a432ba52d707af9df33c7&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYBPJL8Vm14UqzngkD_whSQthgKIBxygrQwHosIdQ5VFqQ&oe=677EC83D&_nc_sid=d885a2
-14,https://instagram.fdel27-1.fna.fbcdn.net/v/t51.2885-15/467577664_18509581051020081_6954526444071406456_n.jpg?stp=dst-jpg_e15_fr_p1080x1080_tt6&_nc_ht=instagram.fdel27-1.fna.fbcdn.net&_nc_cat=109&_nc_ohc=U1rioa0pW6oQ7kNvgHMGWKZ&_nc_gid=c7e77025b4224181a3f65f89ed5ffa14&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYBL_ByEXhyucDaXU9hkedf8NN3oST4cQnSCYt0m_9tF-w&oe=677EBDE6&_nc_sid=d885a2
-15,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/471642941_1331440784721529_3271623926398698799_n.jpg?stp=dst-jpg_e35_p1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=6IPIghw1gk4Q7kNvgERehTV&_nc_gid=a8fbbf9fce1c4053b5efc4a2ce93156f&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYDwf5w8ma24fVnLNeBd0njOJbAL_NlQAGOQDUadDliGCQ&oe=677EA8A8&_nc_sid=d885a2
-16,https://instagram.fdel27-3.fna.fbcdn.net/v/t51.29350-15/472259458_1665395244186013_2233410645634559399_n.jpg?stp=dst-jpg_e35_p1080x1080_tt6&_nc_ht=instagram.fdel27-3.fna.fbcdn.net&_nc_cat=1&_nc_ohc=zqwgl6ga2zIQ7kNvgFFFZWM&_nc_gid=d8740a8df4594a39bcc82b1d2c272a71&edm=ANTKIIoBAAAA&ccb=7-5&oh=00_AYB8R1dlJc88o0hWEJs8G0HeH90CdGrbTJ-k4j1W3eANqw&oe=677EBFBA&_nc_sid=d885a2
